fix bugs; add logging

+++ web-hsdb-import-meta/hearthpwn-parser/src/main/java/ru/bmn/web/hsdb/parser/hearthpwn/UrlContentDiscCache.java
@@ -1,6 +1,9 @@
 package ru.bmn.web.hsdb.parser.hearthpwn;
 
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.core.io.UrlResource;
 import org.springframework.util.DigestUtils;
 
@@ -9,15 +12,17 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.net.URL;
 import java.nio.charset.Charset;
-import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.Map;
 
 /* default */ class UrlContentDiscCache {
-    private static final String PROPERTY_BASE_DIR = "url_content_disc_cache.base_dir";
+	private static final Logger LOG = LogManager.getLogger(UrlContentDiscCache.class);
+
+	private static final String PROPERTY_BASE_DIR = "url_content_disc_cache.base_dir";
 
     private Path baseDir;
     private final Map<String, File> files = new HashMap<>();
@@ -30,11 +35,17 @@ import java.util.Map;
         this.tag = tag;
     }
 
+    /* default */ UrlContentDiscCache(Class<?> clazz) {
+        this.tag = clazz.getSimpleName();
+    }
+
     private void init()
         throws IOException
     {
         if (!this.isInitialized) {
-            String baseDirProperty = System.getProperty(PROPERTY_BASE_DIR);
+        	LOG.info("Init cache...");
+
+            String baseDirProperty = System.getProperty(PROPERTY_BASE_DIR, "");
             if (baseDirProperty.isEmpty()) {
                 throw new IOException(
                     String.format("Cache base dir expected (use %s property)", PROPERTY_BASE_DIR)
@@ -43,52 +54,62 @@ import java.util.Map;
             this.baseDir = Paths.get(baseDirProperty, this.tag);
 
             if (!this.baseDir.toFile().exists()) {
-                Files.createDirectory(this.baseDir);
+            	LOG.info("Create base dir: {}", this.baseDir);
+                Files.createDirectories(this.baseDir);
             }
 
-            try (DirectoryStream<Path> dirStream = Files.newDirectoryStream(this.baseDir)) {
-                for (Path path : dirStream) {
-                    if (path.toFile().isFile()) {
-                        String[] nameParts = path.toFile().getName().split("\\.", 2);
-                        if (nameParts.length != 2) {
-                            throw new RuntimeException(
-                                String.format(
-                                    "Corrupted cache file: %s",
-                                    path.toString()
-                                )
-                            );
-                        }
-                        this.files.put(nameParts[0], path.toFile());
-                    }
+            try {
+            	Iterator<File> fileIterator = FileUtils.iterateFiles(this.baseDir.toFile(), null, true);
+                while (fileIterator.hasNext()) {
+                	File file = fileIterator.next();
+					String[] nameParts = file.getName().split("\\.", 2);
+					if (nameParts.length != 2) {
+						throw new RuntimeException(
+							String.format(
+								"Corrupted cache file: %s",
+								file.toString()
+							)
+						);
+					}
+					this.files.put(nameParts[0], file);
                 }
             }
             finally {
                 this.isInitialized = true;
             }
+            LOG.info("Init completed. Total urls in cache: {}", this.files.entrySet().size());
         }
 
     }
 
-    private void put(URL url)
+    private File put(URL url)
         throws IOException
     {
-        File outputFile = Files.createFile(
-            Paths.get(
-                this.baseDir.toString(),
-                this.urlToFileName(url).toString()
-            )
-        ).toFile();
+    	Path newFilePath = Paths.get(
+			this.baseDir.toString(),
+			this.urlToFileName(url).toString()
+		);
 
-        try (
-            FileOutputStream output = new FileOutputStream(outputFile)
-        ) {
-            IOUtils.copy(
-                new UrlResource(url).getInputStream(),
-                output
-            );
-        }
+    	File outputFile;
+    	if (newFilePath.getParent().toFile().exists() || newFilePath.getParent().toFile().mkdirs()) {
+			outputFile = Files.createFile(newFilePath).toFile();
 
-        this.files.put(this.urlToKey(url), outputFile);
+			try (
+				FileOutputStream output = new FileOutputStream(outputFile)
+			) {
+				IOUtils.copy(
+					new UrlResource(url).getInputStream(),
+					output
+				);
+			}
+
+			this.files.put(this.urlToKey(url), outputFile);
+		}
+		else {
+    		throw new IOException("Can't create cache dir: " + newFilePath.getParent().toString());
+		}
+
+		return outputFile;
     }
 
     private String urlToKey(URL url) {
@@ -101,12 +122,13 @@ import java.util.Map;
             String.format(
                 "%s.%s",
                     this.urlToKey(url),
-                    url.getPath().replaceAll("\\W", "_")
+					String.format("%s__%s", url.getPath(), url.getQuery())
+						.replaceAll("\\W", "_")
             )
         );
     }
 
-    public byte[] get(URL url)
+    private byte[] get(URL url)
         throws IOException
     {
         this.init();
@@ -115,16 +137,11 @@ import java.util.Map;
 
         File file = this.files.get(this.urlToKey(url));
         if (file == null) {
-            this.put(url);
+        	LOG.info("Cache not found: {}", url.toString());
+            file = this.put(url);
         }
-        else {
-            try {
-                result = Files.readAllBytes(file.toPath());
-            }
-            catch (IOException e) {
-                return null;
-            }
-        }
+
+        result = Files.readAllBytes(file.toPath());
         return result;
     }
